Portion of Certificates (#20)
* Portion of kWh for certificates input field

Created a new input field in which a user can specify how many certificates he/she can post for sale.

* Add splitting certificates before posting for sale

If a certificate (that is bound to be post for sale) is larger than the requested Wh the user wants to post for sale, split the certificate into a smaller piece and then post that one for sale

* Use splitAndPublishForSale() method

Replace calling two different methods (2 different transactions)
with only one call to splitAndPublishForSale(). This avoids 2 MetaMask popups when there's a need to split a certificate before publishing it for sale

* Fix split certificates not disappearing

After a certificate is split, it is still shown on the user interface in the Inbox folder. This commit makes the app listen to "logCertificateSplit" event and updates the UI accordingly

* Update dependencies

* Fix Publish for sale modal reappearing

Sometimes the Publish modal would reappear after a close button has been pressed. This commit fixes random reappearances.